Overhaul Windows build system, migrate to MSVC/vcpkg, modernize CMake, remove MinGW, upgrade Qt6, standardize dependencies. See docs/features/build_overhaul/WINDOWS_BUILD_SYSTEM_CMAKE_TOOL_MIGRATION.md for full details.

+++ libs/gui/render/DataProcessor.hpp
@@ -68,7 +68,7 @@ public:
     void stopProcessing();
     
     void createCellsFromLiquiditySlice(const struct LiquidityTimeSlice& slice);
-    void createLiquidityCell(const struct LiquidityTimeSlice& slice, double price, double liquidity, bool isBid);
+    void createLiquidityCell(const struct LiquidityTimeSlice& slice, double price, float liquidity, bool isBid);
     QRectF timeSliceToScreenRect(const struct LiquidityTimeSlice& slice, double price) const;
     
     void setPriceResolution(double resolution);
